ban hoan thien giao dien

+++ Project2/QuanLyBanThuoc/src/Model/Ctrl/KhachhangJpaController.java
@@ -5,16 +5,20 @@
  */
 package Model.Ctrl;
 
+import Model.Ctrl.exceptions.IllegalOrphanException;
 import Model.Ctrl.exceptions.NonexistentEntityException;
-import Model.Khachhang;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import Model.Hoadon;
+import Model.Khachhang;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 
 /**
  *
@@ -32,11 +36,29 @@ public class KhachhangJpaController implements Serializable {
     }
 
     public void create(Khachhang khachhang) {
+        if (khachhang.getHoadonCollection() == null) {
+            khachhang.setHoadonCollection(new ArrayList<Hoadon>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Collection<Hoadon> attachedHoadonCollection = new ArrayList<Hoadon>();
+            for (Hoadon hoadonCollectionHoadonToAttach : khachhang.getHoadonCollection()) {
+                hoadonCollectionHoadonToAttach = em.getReference(hoadonCollectionHoadonToAttach.getClass(), hoadonCollectionHoadonToAttach.getMaHD());
+                attachedHoadonCollection.add(hoadonCollectionHoadonToAttach);
+            }
+            khachhang.setHoadonCollection(attachedHoadonCollection);
             em.persist(khachhang);
+            for (Hoadon hoadonCollectionHoadon : khachhang.getHoadonCollection()) {
+                Khachhang oldMaKHOfHoadonCollectionHoadon = hoadonCollectionHoadon.getMaKH();
+                hoadonCollectionHoadon.setMaKH(khachhang);
+                hoadonCollectionHoadon = em.merge(hoadonCollectionHoadon);
+                if (oldMaKHOfHoadonCollectionHoadon != null) {
+                    oldMaKHOfHoadonCollectionHoadon.getHoadonCollection().remove(hoadonCollectionHoadon);
+                    oldMaKHOfHoadonCollectionHoadon = em.merge(oldMaKHOfHoadonCollectionHoadon);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -45,12 +67,45 @@ public class KhachhangJpaController implements Serializable {
         }
     }
 
-    public void edit(Khachhang khachhang) throws NonexistentEntityException, Exception {
+    public void edit(Khachhang khachhang) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Khachhang persistentKhachhang = em.find(Khachhang.class, khachhang.getMaKH());
+            Collection<Hoadon> hoadonCollectionOld = persistentKhachhang.getHoadonCollection();
+            Collection<Hoadon> hoadonCollectionNew = khachhang.getHoadonCollection();
+            List<String> illegalOrphanMessages = null;
+            for (Hoadon hoadonCollectionOldHoadon : hoadonCollectionOld) {
+                if (!hoadonCollectionNew.contains(hoadonCollectionOldHoadon)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Hoadon " + hoadonCollectionOldHoadon + " since its maKH field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            Collection<Hoadon> attachedHoadonCollectionNew = new ArrayList<Hoadon>();
+            for (Hoadon hoadonCollectionNewHoadonToAttach : hoadonCollectionNew) {
+                hoadonCollectionNewHoadonToAttach = em.getReference(hoadonCollectionNewHoadonToAttach.getClass(), hoadonCollectionNewHoadonToAttach.getMaHD());
+                attachedHoadonCollectionNew.add(hoadonCollectionNewHoadonToAttach);
+            }
+            hoadonCollectionNew = attachedHoadonCollectionNew;
+            khachhang.setHoadonCollection(hoadonCollectionNew);
             khachhang = em.merge(khachhang);
+            for (Hoadon hoadonCollectionNewHoadon : hoadonCollectionNew) {
+                if (!hoadonCollectionOld.contains(hoadonCollectionNewHoadon)) {
+                    Khachhang oldMaKHOfHoadonCollectionNewHoadon = hoadonCollectionNewHoadon.getMaKH();
+                    hoadonCollectionNewHoadon.setMaKH(khachhang);
+                    hoadonCollectionNewHoadon = em.merge(hoadonCollectionNewHoadon);
+                    if (oldMaKHOfHoadonCollectionNewHoadon != null && !oldMaKHOfHoadonCollectionNewHoadon.equals(khachhang)) {
+                        oldMaKHOfHoadonCollectionNewHoadon.getHoadonCollection().remove(hoadonCollectionNewHoadon);
+                        oldMaKHOfHoadonCollectionNewHoadon = em.merge(oldMaKHOfHoadonCollectionNewHoadon);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -68,7 +123,7 @@ public class KhachhangJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -79,6 +134,17 @@ public class KhachhangJpaController implements Serializable {
                 khachhang.getMaKH();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The khachhang with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            Collection<Hoadon> hoadonCollectionOrphanCheck = khachhang.getHoadonCollection();
+            for (Hoadon hoadonCollectionOrphanCheckHoadon : hoadonCollectionOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Khachhang (" + khachhang + ") cannot be destroyed since the Hoadon " + hoadonCollectionOrphanCheckHoadon + " in its hoadonCollection field has a non-nullable maKH field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(khachhang);
             em.getTransaction().commit();
