+++ Project2/QuanLyBanThuoc/src/Model/Ctrl/KhothuocJpaController.java
@@ -5,16 +5,21 @@
  */
 package Model.Ctrl;
 
-import Model.Ctrl.exceptions.NonexistentEntityException;
-import Model.Khothuoc;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import Model.Nhomthuoc;
+import Model.Chitiethoadon;
+import Model.Ctrl.exceptions.IllegalOrphanException;
+import Model.Ctrl.exceptions.NonexistentEntityException;
+import Model.Khothuoc;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 
 /**
  *
@@ -32,11 +37,38 @@ public class KhothuocJpaController implements Serializable {
     }
 
     public void create(Khothuoc khothuoc) {
+        if (khothuoc.getChitiethoadonCollection() == null) {
+            khothuoc.setChitiethoadonCollection(new ArrayList<Chitiethoadon>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Nhomthuoc maNhomThuoc = khothuoc.getMaNhomThuoc();
+            if (maNhomThuoc != null) {
+                maNhomThuoc = em.getReference(maNhomThuoc.getClass(), maNhomThuoc.getMaNhomThuoc());
+                khothuoc.setMaNhomThuoc(maNhomThuoc);
+            }
+            Collection<Chitiethoadon> attachedChitiethoadonCollection = new ArrayList<Chitiethoadon>();
+            for (Chitiethoadon chitiethoadonCollectionChitiethoadonToAttach : khothuoc.getChitiethoadonCollection()) {
+                chitiethoadonCollectionChitiethoadonToAttach = em.getReference(chitiethoadonCollectionChitiethoadonToAttach.getClass(), chitiethoadonCollectionChitiethoadonToAttach.getChitiethoadonPK());
+                attachedChitiethoadonCollection.add(chitiethoadonCollectionChitiethoadonToAttach);
+            }
+            khothuoc.setChitiethoadonCollection(attachedChitiethoadonCollection);
             em.persist(khothuoc);
+            if (maNhomThuoc != null) {
+                maNhomThuoc.getKhothuocCollection().add(khothuoc);
+                maNhomThuoc = em.merge(maNhomThuoc);
+            }
+            for (Chitiethoadon chitiethoadonCollectionChitiethoadon : khothuoc.getChitiethoadonCollection()) {
+                Khothuoc oldKhothuocOfChitiethoadonCollectionChitiethoadon = chitiethoadonCollectionChitiethoadon.getKhothuoc();
+                chitiethoadonCollectionChitiethoadon.setKhothuoc(khothuoc);
+                chitiethoadonCollectionChitiethoadon = em.merge(chitiethoadonCollectionChitiethoadon);
+                if (oldKhothuocOfChitiethoadonCollectionChitiethoadon != null) {
+                    oldKhothuocOfChitiethoadonCollectionChitiethoadon.getChitiethoadonCollection().remove(chitiethoadonCollectionChitiethoadon);
+                    oldKhothuocOfChitiethoadonCollectionChitiethoadon = em.merge(oldKhothuocOfChitiethoadonCollectionChitiethoadon);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -45,12 +77,59 @@ public class KhothuocJpaController implements Serializable {
         }
     }
 
-    public void edit(Khothuoc khothuoc) throws NonexistentEntityException, Exception {
+    public void edit(Khothuoc khothuoc) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Khothuoc persistentKhothuoc = em.find(Khothuoc.class, khothuoc.getMaThuoc());
+            Nhomthuoc maNhomThuocOld = persistentKhothuoc.getMaNhomThuoc();
+            Nhomthuoc maNhomThuocNew = khothuoc.getMaNhomThuoc();
+            Collection<Chitiethoadon> chitiethoadonCollectionOld = persistentKhothuoc.getChitiethoadonCollection();
+            Collection<Chitiethoadon> chitiethoadonCollectionNew = khothuoc.getChitiethoadonCollection();
+            List<String> illegalOrphanMessages = null;
+            for (Chitiethoadon chitiethoadonCollectionOldChitiethoadon : chitiethoadonCollectionOld) {
+                if (!chitiethoadonCollectionNew.contains(chitiethoadonCollectionOldChitiethoadon)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Chitiethoadon " + chitiethoadonCollectionOldChitiethoadon + " since its khothuoc field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            if (maNhomThuocNew != null) {
+                maNhomThuocNew = em.getReference(maNhomThuocNew.getClass(), maNhomThuocNew.getMaNhomThuoc());
+                khothuoc.setMaNhomThuoc(maNhomThuocNew);
+            }
+            Collection<Chitiethoadon> attachedChitiethoadonCollectionNew = new ArrayList<Chitiethoadon>();
+            for (Chitiethoadon chitiethoadonCollectionNewChitiethoadonToAttach : chitiethoadonCollectionNew) {
+                chitiethoadonCollectionNewChitiethoadonToAttach = em.getReference(chitiethoadonCollectionNewChitiethoadonToAttach.getClass(), chitiethoadonCollectionNewChitiethoadonToAttach.getChitiethoadonPK());
+                attachedChitiethoadonCollectionNew.add(chitiethoadonCollectionNewChitiethoadonToAttach);
+            }
+            chitiethoadonCollectionNew = attachedChitiethoadonCollectionNew;
+            khothuoc.setChitiethoadonCollection(chitiethoadonCollectionNew);
             khothuoc = em.merge(khothuoc);
+            if (maNhomThuocOld != null && !maNhomThuocOld.equals(maNhomThuocNew)) {
+                maNhomThuocOld.getKhothuocCollection().remove(khothuoc);
+                maNhomThuocOld = em.merge(maNhomThuocOld);
+            }
+            if (maNhomThuocNew != null && !maNhomThuocNew.equals(maNhomThuocOld)) {
+                maNhomThuocNew.getKhothuocCollection().add(khothuoc);
+                maNhomThuocNew = em.merge(maNhomThuocNew);
+            }
+            for (Chitiethoadon chitiethoadonCollectionNewChitiethoadon : chitiethoadonCollectionNew) {
+                if (!chitiethoadonCollectionOld.contains(chitiethoadonCollectionNewChitiethoadon)) {
+                    Khothuoc oldKhothuocOfChitiethoadonCollectionNewChitiethoadon = chitiethoadonCollectionNewChitiethoadon.getKhothuoc();
+                    chitiethoadonCollectionNewChitiethoadon.setKhothuoc(khothuoc);
+                    chitiethoadonCollectionNewChitiethoadon = em.merge(chitiethoadonCollectionNewChitiethoadon);
+                    if (oldKhothuocOfChitiethoadonCollectionNewChitiethoadon != null && !oldKhothuocOfChitiethoadonCollectionNewChitiethoadon.equals(khothuoc)) {
+                        oldKhothuocOfChitiethoadonCollectionNewChitiethoadon.getChitiethoadonCollection().remove(chitiethoadonCollectionNewChitiethoadon);
+                        oldKhothuocOfChitiethoadonCollectionNewChitiethoadon = em.merge(oldKhothuocOfChitiethoadonCollectionNewChitiethoadon);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -68,7 +147,7 @@ public class KhothuocJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -79,6 +158,22 @@ public class KhothuocJpaController implements Serializable {
                 khothuoc.getMaThuoc();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The khothuoc with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            Collection<Chitiethoadon> chitiethoadonCollectionOrphanCheck = khothuoc.getChitiethoadonCollection();
+            for (Chitiethoadon chitiethoadonCollectionOrphanCheckChitiethoadon : chitiethoadonCollectionOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Khothuoc (" + khothuoc + ") cannot be destroyed since the Chitiethoadon " + chitiethoadonCollectionOrphanCheckChitiethoadon + " in its chitiethoadonCollection field has a non-nullable khothuoc field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            Nhomthuoc maNhomThuoc = khothuoc.getMaNhomThuoc();
+            if (maNhomThuoc != null) {
+                maNhomThuoc.getKhothuocCollection().remove(khothuoc);
+                maNhomThuoc = em.merge(maNhomThuoc);
             }
             em.remove(khothuoc);
             em.getTransaction().commit();
