+++ Project2/QuanLyBanThuoc/src/Model/Ctrl/NhanvienJpaController.java
@@ -5,16 +5,20 @@
  */
 package Model.Ctrl;
 
+import Model.Ctrl.exceptions.IllegalOrphanException;
 import Model.Ctrl.exceptions.NonexistentEntityException;
-import Model.Nhanvien;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import Model.Hoadon;
+import Model.Nhanvien;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 
 /**
  *
@@ -32,11 +36,29 @@ public class NhanvienJpaController implements Serializable {
     }
 
     public void create(Nhanvien nhanvien) {
+        if (nhanvien.getHoadonCollection() == null) {
+            nhanvien.setHoadonCollection(new ArrayList<Hoadon>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Collection<Hoadon> attachedHoadonCollection = new ArrayList<Hoadon>();
+            for (Hoadon hoadonCollectionHoadonToAttach : nhanvien.getHoadonCollection()) {
+                hoadonCollectionHoadonToAttach = em.getReference(hoadonCollectionHoadonToAttach.getClass(), hoadonCollectionHoadonToAttach.getMaHD());
+                attachedHoadonCollection.add(hoadonCollectionHoadonToAttach);
+            }
+            nhanvien.setHoadonCollection(attachedHoadonCollection);
             em.persist(nhanvien);
+            for (Hoadon hoadonCollectionHoadon : nhanvien.getHoadonCollection()) {
+                Nhanvien oldMaNVOfHoadonCollectionHoadon = hoadonCollectionHoadon.getMaNV();
+                hoadonCollectionHoadon.setMaNV(nhanvien);
+                hoadonCollectionHoadon = em.merge(hoadonCollectionHoadon);
+                if (oldMaNVOfHoadonCollectionHoadon != null) {
+                    oldMaNVOfHoadonCollectionHoadon.getHoadonCollection().remove(hoadonCollectionHoadon);
+                    oldMaNVOfHoadonCollectionHoadon = em.merge(oldMaNVOfHoadonCollectionHoadon);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -45,12 +67,45 @@ public class NhanvienJpaController implements Serializable {
         }
     }
 
-    public void edit(Nhanvien nhanvien) throws NonexistentEntityException, Exception {
+    public void edit(Nhanvien nhanvien) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Nhanvien persistentNhanvien = em.find(Nhanvien.class, nhanvien.getMaNV());
+            Collection<Hoadon> hoadonCollectionOld = persistentNhanvien.getHoadonCollection();
+            Collection<Hoadon> hoadonCollectionNew = nhanvien.getHoadonCollection();
+            List<String> illegalOrphanMessages = null;
+            for (Hoadon hoadonCollectionOldHoadon : hoadonCollectionOld) {
+                if (!hoadonCollectionNew.contains(hoadonCollectionOldHoadon)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Hoadon " + hoadonCollectionOldHoadon + " since its maNV field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            Collection<Hoadon> attachedHoadonCollectionNew = new ArrayList<Hoadon>();
+            for (Hoadon hoadonCollectionNewHoadonToAttach : hoadonCollectionNew) {
+                hoadonCollectionNewHoadonToAttach = em.getReference(hoadonCollectionNewHoadonToAttach.getClass(), hoadonCollectionNewHoadonToAttach.getMaHD());
+                attachedHoadonCollectionNew.add(hoadonCollectionNewHoadonToAttach);
+            }
+            hoadonCollectionNew = attachedHoadonCollectionNew;
+            nhanvien.setHoadonCollection(hoadonCollectionNew);
             nhanvien = em.merge(nhanvien);
+            for (Hoadon hoadonCollectionNewHoadon : hoadonCollectionNew) {
+                if (!hoadonCollectionOld.contains(hoadonCollectionNewHoadon)) {
+                    Nhanvien oldMaNVOfHoadonCollectionNewHoadon = hoadonCollectionNewHoadon.getMaNV();
+                    hoadonCollectionNewHoadon.setMaNV(nhanvien);
+                    hoadonCollectionNewHoadon = em.merge(hoadonCollectionNewHoadon);
+                    if (oldMaNVOfHoadonCollectionNewHoadon != null && !oldMaNVOfHoadonCollectionNewHoadon.equals(nhanvien)) {
+                        oldMaNVOfHoadonCollectionNewHoadon.getHoadonCollection().remove(hoadonCollectionNewHoadon);
+                        oldMaNVOfHoadonCollectionNewHoadon = em.merge(oldMaNVOfHoadonCollectionNewHoadon);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -68,7 +123,7 @@ public class NhanvienJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -79,6 +134,17 @@ public class NhanvienJpaController implements Serializable {
                 nhanvien.getMaNV();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The nhanvien with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            Collection<Hoadon> hoadonCollectionOrphanCheck = nhanvien.getHoadonCollection();
+            for (Hoadon hoadonCollectionOrphanCheckHoadon : hoadonCollectionOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Nhanvien (" + nhanvien + ") cannot be destroyed since the Hoadon " + hoadonCollectionOrphanCheckHoadon + " in its hoadonCollection field has a non-nullable maNV field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(nhanvien);
             em.getTransaction().commit();
